Backend/VolunteerNowProject/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
awscli==1.33.41
awsebcli==3.20.10
Babel==2.15.0
beautifulsoup4==4.12.3
blis==0.7.11
botocore==1.34.159
bs4==0.0.2
cachetools==5.3.3
camelot-py==0.9.0
catalogue==2.0.10
cement==2.8.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==42.0.7
cymem==2.0.8
defusedxml==0.7.1
dialogflow==2.0.0
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
django-filter==24.2
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
docutils==0.16
drf-yasg==1.21.7
et-xmlfile==1.1.0
etils==1.8.0
filelock==3.14.0
flatbuffers==24.3.25
flax==0.8.3
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
google-ai-generativelanguage==0.6.2
google-api-core==1.34.1
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-dialogflow==2.30.0
google-generativeai==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.48.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpie==3.2.3
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
imagesize==1.4.1
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
intel-openmp==2021.4.0
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
keras==3.3.3
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
npm==0.1.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optax==0.2.2
optional-django==0.1.0
optree==0.11.0
orbax-checkpoint==0.5.11
packaging==24.0
pandas==2.2.2
pathspec==0.10.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
pluggy==1.5.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==3.20.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pypiwin32==223
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
razorpay==1.4.2
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.7.2
s3transfer==0.10.2
safetensors==0.4.3
scipy==1.13.0
semantic-version==2.8.5
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
srsly==2.4.8
sympy==1.12
tabula-py==2.9.3
tabulate==0.9.0
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorstore==0.1.59
termcolor==1.1.0
thinc==8.2.3
tokenizers==0.19.1
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0+cpu
torchvision==0.18.0+cpu
tqdm==4.66.4
transformers==4.41.0
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
vboxapi==1.0
wasabi==1.1.2
wcwidth==0.1.9
weasel==0.3.4
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
